+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-There are no changes yet.
+### Added
+
+- Axes and lines now have `class`/`className` prop.
 
 
 ## [0.2.1] - 2018-11-05